backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
django==4.2.4; python_version >= '3.8'
django-ninja==0.22.2; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
psycopg==3.1.10; python_version >= '3.7'
pydantic==1.10.12; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.7.1; python_version < '3.11'
uvicorn==0.23.2; python_version >= '3.8'
